[SG-110] fix: nested Map의 key를 notificationId -> 정렬된 리스트의 index 값으로 변경

+++ core/model/src/main/kotlin/com/captures2024/soongan/core/model/utils/TypeAlias.kt
@@ -2,6 +2,6 @@ package com.captures2024.soongan.core.model.utils
 
 import com.captures2024.soongan.core.model.dto.Notification
 
-typealias NotificationsTable = Map<NotificationType, Map<Long, Notification>>
+typealias NotificationsTable = Map<NotificationType, Map<Int, Notification>>
 
 typealias NotificationsCountTable = Map<NotificationType, Int>
